# perf: optimize state machine for better branch prediction