## Summary

Analyzed assembly output using `cargo asm` and implemented several optimizations to improve branch prediction and reduce instruction count in hot paths.

### Changes

1. **Explicit enum discriminants**: Added explicit `u8` discriminants (0-6) to `State` enum for more efficient integer comparisons at the assembly level

2. **Integer-based state matching**: Changed from enum pattern matching to integer discriminant matching (`*state as u8`) with `unreachable_unchecked()` hint for impossible cases, reducing branch prediction overhead

3. **Fast-path optimization**: Added early-return fast paths in `top()` and `in_string()` functions for the most common case (non-special characters), reducing instruction count in the hot path

### Performance Results

Measured with criterion benchmarks:

| Benchmark | Before | After | Improvement |
|-----------|--------|-------|-------------|
| tsconfig | 2.219µs | 2.129µs | **3.5-4.7%** |
| no_comments | 472ns | 451ns | **4.5-5.4%** |
| minimal_comments | 121ns | 112ns | **6.3-8.2%** |
| large_with_comments | 11.375µs | 11.109µs | **2.5-3.6%** |

### Assembly Improvements

- State transitions now use simpler integer comparisons (`cmp w8, #2`) instead of multiple enum checks
- Better instruction cache locality with more predictable branching
- Fast-path in hot functions reduces conditional branches for common cases

## Test plan

- [x] All 35 existing tests pass
- [x] Benchmark suite shows 2.5-8% performance improvements across all workloads
- [x] Assembly output verified with `cargo asm` shows improved code generation

🤖 Generated with [Claude Code](https://claude.com/claude-code)